fix move issues
